Move layout components to within components
