+++ Source/SelfService/Web/components/layout/layout.scss
