+++ Source/SelfService/Web/components/layout/topNavBar.tsx
@@ -3,7 +3,7 @@
 
 import React from 'react';
 
-import { ShortInfoWithEnvironment } from '../api/solutions/api';
+import { ShortInfoWithEnvironment } from '../../api/solutions/api';
 
 import { Box, Grid } from '@mui/material';
 import { BreadCrumbContainer, BreadcrumbsRoute } from './breadcrumbs';
