+++ Source/SelfService/Web/solutions/backupsScreen.tsx
@@ -11,8 +11,8 @@ import {
 
 import { getApplication, HttpResponseApplication } from '../api/solutions/application';
 import { ViewCard } from '../solutions/backup/viewCard';
-import { getMenuWithApplication, LayoutWithSidebar } from '../layout/layoutWithSidebar';
-import { BreadCrumbContainer } from '../layout/breadcrumbs';
+import { getMenuWithApplication, LayoutWithSidebar } from '../components/layout/layoutWithSidebar';
+import { BreadCrumbContainer } from '../components/layout/breadcrumbs';
 import { useRouteApplicationParams } from '../utils/route';
 import { ListView } from './backup/listView';
 import { useGlobalContext } from '../context/globalContext';
